refactor: refactoring while there is some time)))
